Another script to achieve the same effect
This one is flawless because it moves the cube around. This isn't very
significant but it's pretty. Also, it will encourage me to learn how to
use Quaternion more effectively.

There are built in functions to rotate an object, I choose not to use
them yet without struggling to do it myself first.

Things I should do:
1) Comment on the code. I didn't do it because I don't think it needs a
lot of explanation. Besides, it's like a "Hello World!" example.