2nd commit
This one is flawless because it moves the cube around rather than the
camera. It doesn't run into any funky degree changes therefore no need
for if/else statements which gave me quite a bit of a headache in the
1st commit.

I know there are built in functions to do this, but I first rather
struggle with doing it myself before utilizing those functions.

Things I should have done:
1) Comment my code. I didn't do it here because it's short and simple,
like a sort of "Hello World!" example.